Batchnorm in Makemore
This notebook implements BatchNorm into the Makemore NN framework. This helps counteract the formation of dead neurons and keeps activation values out of the tails of the tanh function. Also included is a rewrite of part of the Pytorch nn module. The performance of Makemore is no longer bottlenecked by optimization, but rather by context length.  Finally there are some useful diagnostic visualization code blocks.